refactor: update UserServices user_impact methods
