+++ app/models/user.rb
@@ -1,4 +1,6 @@
 class User < ApplicationRecord
+  include UserServices::UserImpact
+
   validates :email, uniqueness: { case_sensitive: true }, format: { with: URI::MailTo::EMAIL_REGEXP }
 
   before_validation { email.downcase! }
@@ -10,29 +12,6 @@ class User < ApplicationRecord
 
   delegate :last_donation_at, to: :user_donation_stats
   delegate :can_donate?, to: :user_donation_stats
-
-  def user_impact
-    donation_balances = Graphql::RibonApi::Client.query(Graphql::Queries::FetchDonationBalances::Query)
-
-    user_donations = donation_balances.original_hash['data']['donationBalances'].select do |donation|
-      donation['user'] == hashed_email
-    end
-
-    result = []
-
-    user_donations.each do |donation|
-      result << {
-        non_profit: donation['nonProfit'],
-        total_donated: donation['totalDonated']
-      }
-    end
-
-    result
-  end
-
-  def hashed_email
-    "0x#{Digest::Keccak.new(256).hexdigest(email)}"
-  end
 
   private
